Move jhenner's manual tests.
