+++ cfme/tests/configure/test_schedule_operations.py
@@ -3,6 +3,7 @@ import pytest
 import pytz
 from dateutil import parser, relativedelta
 
+from cfme.utils import conf
 from cfme.base.ui import BaseLoggedInPage
 from cfme.infrastructure.provider.virtualcenter import VMwareProvider
 from cfme.markers.env_markers.provider import ONE
@@ -220,127 +221,3 @@ def test_schedule_timer(appliance, run_types, host_with_credentials, request, cu
                 schedule.last_run_date).strftime('%m%d%H'),
             delay=60, timeout="10m", fail_func=appliance.server.browser.refresh,
             message="Scheduled task didn't run in appropriate time set")
-
-
-@pytest.mark.manual
-@pytest.mark.tier(2)
-def test_schedule_for_all_clusters_vms():
-    """
-    Navigate to add new schedule page(Configuration->Region->Schedules)
-    Fill all required fields
-    Select all vms for clusters in filter
-    Set timer
-    Save changes
-    Result: Task run successfully for selected filter
-
-    Polarion:
-        assignee: jhenner
-        casecomponent: config
-        caseimportance: medium
-        initialEstimate: 1/8h
-    """
-    pass
-
-
-@pytest.mark.manual
-@pytest.mark.tier(2)
-def test_schedule_for_all_host_vms():
-    """
-    Navigate to add new schedule page(Configuration->Region->Schedules)
-    Fill all required fields
-    Select all vms for host in filter
-    Set timer
-    Save changes
-    Result: Task run successfully for selected filter
-
-    Polarion:
-        assignee: jhenner
-        casecomponent: config
-        caseimportance: medium
-        initialEstimate: 1/8h
-    """
-    pass
-
-
-@pytest.mark.manual
-@pytest.mark.tier(2)
-def test_schedule_for_all_providers_vms():
-    """
-    Navigate to add new schedule page(Configuration->Region->Schedules)
-    Fill all required fields
-    Select all vms for provider in filter
-    Set timer
-    Save changes
-    Result: Task run successfully for selected filter
-
-    Polarion:
-        assignee: jhenner
-        casecomponent: config
-        caseimportance: medium
-        initialEstimate: 1/8h
-    """
-    pass
-
-
-@pytest.mark.manual
-@pytest.mark.tier(2)
-def test_schedule_global_filters():
-    """
-    Navigate to add new schedule page(Configuration->Region->Schedules)
-    Fill all required fields
-    Select some global filter
-    Set timer
-    Save changes
-    Result: Task run successfully for selected filter
-
-    Polarion:
-        assignee: jhenner
-        casecomponent: config
-        caseimportance: medium
-        initialEstimate: 1/8h
-    """
-    pass
-
-
-@pytest.mark.manual
-@test_requirements.report
-def test_schedule_send_multiple_recipients():
-    """
-    Using scripts/smtp_collectior.py
-    Pre-requirements:
-    Set CFME Notifier role ON
-    CFME set up to use external SMTP using your local pc ip (port 1025)
-    Allow 1025 on local machine (firewall-cmd --add-port 1025/tcp)
-    Turn on smtp_collector in cfme_tests env
-    Set up schedule to send to 10 contacts.
-    1)Send schedule mail using default From address
-    2)Change From address and resend schedule mail
-    Check http://localhost:1026/messages.html for both sets of mail
-    present.
-
-    Polarion:
-        assignee: jhenner
-        caseimportance: medium
-        initialEstimate: None
-    """
-    pass
-
-
-@pytest.mark.manual
-@pytest.mark.tier(2)
-def test_schedule_compliance_check():
-    """
-    Navigate to add new schedule page(Configuration->Region->Schedules)
-    Fill all required fields
-    Select compliance check in filter
-    Set timer
-    Save changes
-    Result: Task run successfully
-
-    Polarion:
-        assignee: jhenner
-        casecomponent: config
-        caseimportance: medium
-        initialEstimate: 1/8h
-    """
-    pass
