+++ cfme/tests/intelligence/test_dashboard.py
@@ -161,3 +161,70 @@ def test_widgets_reorder(dashboards, soft_assert, request):
     assert previous_names[2:] == new_names[2:]
     assert previous_names[0] == new_names[1]
     assert previous_names[1] == new_names[0]
+
+
+@pytest.mark.manual
+@test_requirements.dashboard
+@pytest.mark.tier(3)
+def test_dashboard_layouts_match():
+    """
+    https://bugzilla.redhat.com/show_bug.cgi?id=1518766
+
+    Polarion:
+        assignee: jhenner
+        casecomponent: web_ui
+        caseimportance: medium
+        initialEstimate: 1/12h
+    """
+    pass
+
+
+@pytest.mark.manual
+@test_requirements.dashboard
+@pytest.mark.tier(3)
+def test_dashboard_widgets_fullscreen():
+    """
+    https://bugzilla.redhat.com/show_bug.cgi?id=1518901
+
+    Polarion:
+        assignee: jhenner
+        casecomponent: web_ui
+        caseimportance: low
+        initialEstimate: 1/12h
+    """
+    pass
+
+
+@pytest.mark.manual
+@test_requirements.dashboard
+@pytest.mark.tier(3)
+def test_identical_dashboard_widgets():
+    """
+    test_identical_dashboard_widgets
+
+    Polarion:
+        assignee: jhenner
+        casecomponent: report
+        caseimportance: low
+        caseposneg: negative
+        initialEstimate: 1/12h
+    """
+    pass
+
+
+@pytest.mark.manual
+@test_requirements.dashboard
+@pytest.mark.tier(3)
+def test_dashboard_chart_widgets_size_in_modal():
+    """
+    Test whether dashboard chart widgets have correct size in modal
+    window.
+
+    Polarion:
+        assignee: jhenner
+        casecomponent: report
+        caseimportance: low
+        initialEstimate: 1/6h
+        testtype: nonfunctional
+    """
+    pass
